requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
PyGithub==1.25.2
PyYAML==3.11
SQLAlchemy==0.9.8
Werkzeug==0.10.4
argparse==1.2.1
itsdangerous==0.24
redis==2.10.3
wsgiref==0.1.2
